crypto_api/requirements.txt
fastapi[all]
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
pytest
python-dotenv
httpx
celery[redis]
redis
